mainstuff/requirements.txt
alembic==1.5.6
cffi==1.14.5
click==8.0.3
cryptography==3.4.6
Flask==2.0.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.4
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
protobuf==3.15.5
psycopg2-binary
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==2.0.2
WTForms==2.3.3